Split B3Propagator functionality to two helper classes B3PropagatorExtractor & B3PropagatorInjector
